requirements.txt
streamlit>=1.36.0
gspread>=6.0.0
google-auth>=2.0.0
pandas
